+++ apps/docs/app/component/[slug]/layout.client.tsx
@@ -224,6 +224,24 @@ export function ComponentLayoutClient({
           <Tabs slug={component.slug} />
 
           <Core.GdsFlex flex-direction="column" id="component-content" gap="xl">
+            {component.slug === 'icon' && (
+              <Core.GdsAlert
+                variant="notice"
+                buttonLabel="View all icons"
+                onClick={() => router.push('/studio/icons')}
+              >
+                <Core.GdsText
+                  font="body-book-m"
+                  display="inline"
+                  font-weight="bold"
+                >
+                  Explore our icon library!
+                </Core.GdsText>
+                {` `}
+                Discover our searchable icon library with regular and solid
+                variants.
+              </Core.GdsAlert>
+            )}
             {section === 'overview' && (component.preamble || anatomyImage) && (
               <Core.GdsFlex
                 flex-direction="column"
@@ -270,7 +288,9 @@ export function ComponentLayoutClient({
                 )}
               </Core.GdsFlex>
             )}
+
             {children}
+
             {component.soon && (
               <Core.GdsAlert variant="notice">
                 {component.title} documentation is in progress.
@@ -316,35 +336,6 @@ export function ComponentLayoutClient({
               section={section}
               versus={versus}
             />
-
-            {/* {component.platform?.web && (
-              <Core.GdsFlex gap="xs" flex-direction="column" padding="m">
-                <Core.GdsDiv height="1px" background="primary"></Core.GdsDiv>
-                <Core.GdsText tag="small" color="neutral-02">
-                  Platform
-                </Core.GdsText>
-                <Core.GdsFlex gap="s">
-                  {component.platform.web && (
-                    <Core.GdsFlex align-items="center" gap="4xs">
-                      <Core.IconCompassRound></Core.IconCompassRound>
-                      Web
-                    </Core.GdsFlex>
-                  )}
-                  {component.platform.ios && (
-                    <Core.GdsFlex align-items="center" gap="4xs">
-                      <Core.IconPhoneDynamicIsland></Core.IconPhoneDynamicIsland>
-                      iOS
-                    </Core.GdsFlex>
-                  )}
-                  {component.platform.android && (
-                    <Core.GdsFlex align-items="center" gap="4xs">
-                      <Core.IconRobot></Core.IconRobot>
-                      Android
-                    </Core.GdsFlex>
-                  )}
-                </Core.GdsFlex>
-              </Core.GdsFlex>
-            )} */}
           </Core.GdsFlex>
         )}
       </Core.GdsGrid>