+++ libs/core/src/components/context-menu/context-menu.component.ts
@@ -22,7 +22,6 @@ import {
 import { GdsButton } from '../button/button.component'
 import { IconDotGridOneHorizontal } from '../icon/icons/dot-grid-one-horizontal'
 import { GdsPopover } from '../popover/popover.component'
-import ContextMenuStyles from './context-menu.styles'
 
 export { GdsMenuItem }
 export { GdsMenuHeading } from '../../primitives/menu/menu-heading.component'
@@ -52,7 +51,7 @@ export { GdsMenuHeading } from '../../primitives/menu/menu-heading.component'
 export class GdsContextMenu extends withMarginProps(
   withLayoutChildProps(GdsElement),
 ) {
-  static styles = [tokens, ContextMenuStyles]
+  static styles = [tokens]
 
   static shadowRootOptions: ShadowRootInit = {
     mode: 'open',