+++ libs/core/src/components/icon/icons/circle-ban.ts
@@ -1,0 +1,5 @@
+import { IconCircleBan } from './circle-ban.component'
+
+IconCircleBan.define()
+
+export { IconCircleBan }