+++ libs/core/src/components/icon/icons/shapes.ts
@@ -1,0 +1,5 @@
+import { IconShapes } from './shapes.component'
+
+IconShapes.define()
+
+export { IconShapes }