+++ libs/react/src/lib/form/text/text.tsx
@@ -4,6 +4,9 @@ interface TextProps {
   children: ReactNode
 }
 
+/**
+ * @deprecated Please use the `gds-text` web component from green-core instead
+ */
 export const Text = ({ children }: TextProps) => (
   <span className="form-text">{children}</span>
 )