+++ libs/react/src/lib/icon/icons/IconBulletList.tsx
@@ -1,0 +1,1 @@
+export { IconBulletList } from '@sebgroup/green-core/react';