+++ libs/react/src/lib/icon/icons/IconCircleDots.tsx
@@ -1,0 +1,1 @@
+export { IconCircleDots } from '@sebgroup/green-core/react';