+++ libs/react/src/lib/icon/icons/IconEmojiNeutral.tsx
@@ -1,0 +1,1 @@
+export { IconEmojiNeutral } from '@sebgroup/green-core/react';